requirements.txt
beautifulsoup4==4.12.3
Pillow==10.4.0
PyQt5==5.15.11
Requests==2.32.3
